+++ spiel/__init__.py
@@ -1,3 +1,4 @@
 from spiel.constants import __version__
-from spiel.deck import Deck, Slide
+from spiel.deck import Deck
+from spiel.slide import Slide
 from spiel.triggers import Triggers
