+++ tests/test_slide.py
@@ -1,0 +1,5 @@
+from spiel import Slide, Triggers
+
+
+def test_can_render_default_slide() -> None:
+    Slide().render(triggers=Triggers.new())
